backend-and-dash/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
paho-mqtt==2.1.0
aiofiles==24.1.0
PyNaCl==1.5.0
httpx==0.28.1
pytest==8.3.4
